+++ pkg/distributor/distributor_test.go
@@ -559,9 +559,7 @@ func TestDistributor_MetricsCleanup(t *testing.T) {
 		cortex_distributor_ingester_query_failures_total{ingester="ingester-1"} 1
 		`), permanentMetrics...))
 
-	err = testutil.GatherAndCompare(reg, strings.NewReader(""), removedMetrics...)
-	require.ErrorContains(t, err, "expected metric name(s) not found")
-	require.ErrorContains(t, err, strings.Join(removedMetrics, " "))
+	require.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(""), removedMetrics...))
 }
 
 func TestDistributor_PushIngestionRateLimiter(t *testing.T) {
@@ -2211,7 +2209,7 @@ func BenchmarkDistributor_Push(b *testing.B) {
 			require.NoError(b, err)
 
 			// Start the distributor.
-			distributor, err := New(distributorCfg, clientConfig, overrides, ingestersRing, true, nil, log.NewNopLogger())
+			distributor, err := New(distributorCfg, clientConfig, overrides, ingestersRing, true, prometheus.NewRegistry(), log.NewNopLogger())
 			require.NoError(b, err)
 			require.NoError(b, services.StartAndAwaitRunning(context.Background(), distributor))
 
@@ -3315,6 +3313,10 @@ type noopIngester struct {
 
 func (i *noopIngester) Close() error {
 	return nil
+}
+
+func (i *noopIngester) PushPreAlloc(ctx context.Context, in *cortexpb.PreallocWriteRequest, opts ...grpc.CallOption) (*cortexpb.WriteResponse, error) {
+	return nil, nil
 }
 
 func (i *noopIngester) Push(ctx context.Context, req *cortexpb.WriteRequest, opts ...grpc.CallOption) (*cortexpb.WriteResponse, error) {